[Refactor]: package 구조 개선
# Conflicts:
#	src/main/java/com/backend/domain/cash/controller/ApiV1CashController.java
