+++ src/main/java/com/backend/domain/bid/service/BidService.java
@@ -11,8 +11,8 @@ import com.backend.domain.product.entity.Product;
 import com.backend.domain.product.enums.AuctionStatus;
 import com.backend.global.exception.ServiceException;
 import com.backend.domain.notification.service.BidNotificationService;
-import com.backend.global.rsData.RsData;
-import com.backend.global.webSocket.service.WebSocketService;
+import com.backend.global.response.RsData;
+import com.backend.global.websocket.service.WebSocketService;
 import jakarta.persistence.EntityManager;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
