+++ src/main/java/com/backend/domain/member/service/MemberService.java
@@ -8,9 +8,9 @@ import com.backend.domain.member.dto.MemberSignUpResponseDto;
 import com.backend.domain.member.entity.Member;
 import com.backend.domain.member.repository.MemberRepository;
 import com.backend.global.exception.ServiceException;
-import com.backend.global.rsData.RsData;
+import com.backend.global.response.RsData;
 import com.backend.global.security.JwtUtil;
-import com.backend.global.util.RedisUtil;
+import com.backend.global.redis.RedisUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
