+++ src/main/java/com/backend/domain/product/controller/ApiV1ProductController.java
@@ -9,8 +9,8 @@ import com.backend.domain.product.enums.ProductSearchSortType;
 import com.backend.domain.product.enums.SaleStatus;
 import com.backend.domain.product.service.ProductService;
 import com.backend.global.exception.ServiceException;
-import com.backend.global.rsData.RsData;
-import com.backend.standard.page.dto.PageDto;
+import com.backend.global.response.RsData;
+import com.backend.global.page.dto.PageDto;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
