+++ src/main/java/com/backend/domain/product/controller/ApiV1ProductControllerDocs.java
@@ -4,8 +4,8 @@ import com.backend.domain.product.dto.*;
 import com.backend.domain.product.enums.AuctionStatus;
 import com.backend.domain.product.enums.ProductSearchSortType;
 import com.backend.domain.product.enums.SaleStatus;
-import com.backend.global.rsData.RsData;
-import com.backend.standard.page.dto.PageDto;
+import com.backend.global.response.RsData;
+import com.backend.global.page.dto.PageDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
