+++ src/main/java/com/backend/global/page/dto/PageableDto.java
@@ -1,0 +1,10 @@
+package com.backend.global.page.dto;
+
+public record PageableDto(
+        int currentPage,
+        int pageSize,
+        int totalPages,
+        long totalElements,
+        boolean hasNext,
+        boolean hasPrevious
+) {}
