--- src/main/java/com/backend/global/rsData/RsData.java
@@ -1,19 +1,0 @@
-package com.backend.global.rsData;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import org.springframework.lang.NonNull;
-
-public record RsData<T>(
-        @NonNull String resultCode,
-        @JsonIgnore int statusCode,
-        @NonNull String msg,
-        @NonNull T data
-) {
-    public RsData(String resultCode, String msg) {
-        this(resultCode, msg, null);
-    }
-
-    public RsData(String resultCode, String msg, T data) {
-        this(resultCode, Integer.parseInt(resultCode.split("-", 2)[0]), msg, data);
-    }
-}
