+++ src/main/java/com/backend/global/security/JwtAuthenticationFilter.java
@@ -2,7 +2,7 @@ package com.backend.global.security;
 
 import com.backend.domain.member.entity.Member;
 import com.backend.domain.member.repository.MemberRepository;
-import com.backend.global.util.RedisUtil;
+import com.backend.global.redis.RedisUtil;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
