--- src/main/java/com/backend/global/springDoc/SpringDocConfig.java
@@ -1,21 +1,0 @@
-package com.backend.global.springDoc;
-
-import io.swagger.v3.oas.annotations.OpenAPIDefinition;
-import io.swagger.v3.oas.annotations.enums.SecuritySchemeType;
-import io.swagger.v3.oas.annotations.info.Info;
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
-import io.swagger.v3.oas.annotations.security.SecurityScheme;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@OpenAPIDefinition(info = @Info(title = "실시간 경매 API", version = "beta", description = "BE 실시간 경매 API 서버 문서입니다."),
-        security = { @SecurityRequirement(name = "bearerAuth") } // Authorize 버튼 활성 + 모든 API에 적용
-)
-@SecurityScheme(
-        name = "bearerAuth",
-        type = SecuritySchemeType.HTTP,
-        scheme = "bearer",
-        bearerFormat = "JWT"
-)
-public class SpringDocConfig {
-}
