--- src/main/java/com/backend/global/util/RedisUtil.java
@@ -1,26 +1,0 @@
-package com.backend.global.util;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.stereotype.Component;
-
-import java.util.concurrent.TimeUnit;
-
-@Component
-@RequiredArgsConstructor
-public class RedisUtil {
-
-    private final RedisTemplate<String, Object> redisTemplate;
-
-    public void setData(String key, String value, Long expiredTime) {
-        redisTemplate.opsForValue().set(key, value, expiredTime, TimeUnit.MILLISECONDS);
-    }
-
-    public Object getData(String key) {
-        return redisTemplate.opsForValue().get(key);
-    }
-
-    public void deleteData(String key) {
-        redisTemplate.delete(key);
-    }
-}
