--- src/main/java/com/backend/global/webSocket/config/WebSocketConfig.java
@@ -1,29 +1,0 @@
-package com.backend.global.webSocket.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.messaging.simp.config.MessageBrokerRegistry;
-import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
-import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
-import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerConfigurer;
-
-@Configuration
-@EnableWebSocketMessageBroker
-public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
-
-    @Override
-    public void registerStompEndpoints(StompEndpointRegistry registry) {
-        // WebSocket 연결 엔드포인트 설정
-        registry.addEndpoint("/ws")
-                .setAllowedOriginPatterns("*") // CORS 설정
-                .withSockJS(); // SockJS 지원
-    }
-
-    @Override
-    public void configureMessageBroker(MessageBrokerRegistry registry) {
-        // 클라이언트가 구독할 경로
-        registry.enableSimpleBroker("/topic");
-        // 클라이언트에서 서버로 메시지를 보낼 때 사용할 경로
-        registry.setApplicationDestinationPrefixes("/app");
-
-    }
-}
