--- src/main/java/com/backend/standard/page/dto/PageableDto.java
@@ -1,10 +1,0 @@
-package com.backend.standard.page.dto;
-
-public record PageableDto(
-        int currentPage,
-        int pageSize,
-        int totalPages,
-        long totalElements,
-        boolean hasNext,
-        boolean hasPrevious
-) {}
